requirements.txt
django
pandas
imgkit
pillow
opencv-python
pandas_profiling

